requirements.txt
streamlit
google-generativeai
pandas
matplotlib
wordcloud
requests
beautifulsoup4
PyPDF2 